Update entire chapter 15 in Section 03

+++ sec-03-oop/15-virtual-functions/cpp/base_pointers_references_to_derived.cpp
@@ -27,7 +27,7 @@ public:
     }
 
     std::string_view get_name() const { return "Derived"; }
-    int get_value() const { return value; }
+    double get_value() const { return value; }
 };
 
 int main()
